+++ Nos/Views/ExpandingTextFieldAndSubmitButton.swift
@@ -12,9 +12,9 @@ struct ExpandingTextFieldAndSubmitButton: View {
     
     var placeholder: String
     @Binding var reply: String
+    var focus: FocusState<Bool>.Binding
     var action: () async -> Void
     
-    @FocusState private var textEditorInFocus
     @State private var showPostButton = false
     @State var disabled = false
     
@@ -35,7 +35,7 @@ struct ExpandingTextFieldAndSubmitButton: View {
                 .background(Color.appBg)
                 .cornerRadius(17.5)
                 .frame(maxHeight: 270)
-                .focused($textEditorInFocus)
+                .focused(focus)
             
             if showPostButton {
                 Button(
@@ -51,14 +51,11 @@ struct ExpandingTextFieldAndSubmitButton: View {
                         Localized.post.view
                     }
                 )
-                .transition(.move(edge: .trailing))
                 .disabled(disabled)
             }
         }
-        .onChange(of: textEditorInFocus) { bool in
-            withAnimation(.spring(response: 0.2)) {
-                showPostButton = bool
-            }
+        .onChange(of: focus.wrappedValue) { bool in
+            showPostButton = bool
         }
         .padding(8)
     }