+++ Nos/Views/HomeFeedView.swift
@@ -107,6 +107,9 @@ struct HomeFeedView: View {
             .navigationDestination(for: Event.self) { note in
                 RepliesView(note: note)
             }
+            .navigationDestination(for: ReplyToNavigationDestination.self) { destination in 
+                RepliesView(note: destination.note, showKeyboard: true)
+            }
             .navigationDestination(for: Author.self) { author in
                 if router.currentPath.wrappedValue.count == 1 {
                     ProfileView(author: author)