+++ Nos/Views/NotificationsView.swift
@@ -53,6 +53,9 @@ struct NotificationsView: View {
             .navigationDestination(for: Event.self) { note in
                 RepliesView(note: note)
             }
+            .navigationDestination(for: ReplyToNavigationDestination.self) { destination in 
+                RepliesView(note: destination.note, showKeyboard: true)
+            }
             .navigationDestination(for: Author.self) { author in
                 ProfileView(author: author)
             }