+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Made the reply button on notes easier to tap, and it now presents the keyboard when tapped.
 - Added support for reposting notes.
 - Fixed a bug where you could post the same reply multiple times.
 - Fixed a bug where the user's follow list could be erased on the first launch after importing a new key.